Open Parser class so that applications can add their own parameter types
Refs #14200
